% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-trajectory.R, R/roxygen-test.R
\name{anyNameButFunctionNameIsUnique}
\alias{anyNameButFunctionNameIsUnique}
\alias{prepare_trajectory}
\alias{standardise_names}
\alias{alsoTheQuestion}
\alias{alsoToBar}
\alias{alsoNotToBar}
\title{Utility functions to prepare study trajectory}
\usage{
prepare_trajectory(.q_trj)

standardise_names(.q_trj)

alsoTheQuestion(z)

alsoToBar(x)

alsoNotToBar(x, y)

alsoTheQuestion(z)
}
\arguments{
\item{.q_trj}{trajectory extracted by Q with OSN and ECTRL add other more often used params here}

\item{z}{a param just for theQuestion.

Could put here, but easiest if all params are
described in the same place, with page documentation, so none get repeated.}

\item{x}{a param for toBar and notToBar}

\item{y}{a param just for notToBar}
}
\value{
Hard to have one return section for >>>> ALL <<<<< functions,

Hard to have one return section for all functions,
might want to have a manual list here.
}
\description{
some descripion

Group of functions Description section
}
\details{
more details.

Group of functions Details paragraph.
}
\section{Another section}{



Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}

\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

