% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqutils-zip.R
\name{read_csv12}
\alias{read_csv12}
\title{Read csv files switching csv and csv2}
\usage{
read_csv12(.fn, .colspec = NULL, .show_col_types = FALSE, ...)
}
\arguments{
\item{.fn}{filename (including file path)}

\item{.colspec}{optional reading of selective columns}

\item{...}{}
}
\value{
df return data frame
}
\description{
Utility function wrapper for readr::read_csv and readr::read_csv2.
Function tests the csv file and then picks the respective parser csv or csv2.
}
\examples{
\dontrun{
  read_csv12(file_name)
}
}
