% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqutils-zip.R
\name{read_zip}
\alias{read_zip}
\title{Read zip-archive}
\usage{
read_zip(
  .pth,
  .archive,
  .files = NULL,
  .path_unzip = "tmp_unzip",
  .show_col_types = FALSE,
  ...
)
}
\arguments{
\item{.pth}{path to zip-archive}

\item{.archive}{actual zip-file}

\item{.files}{vector of filenames in zip-archive}

\item{.path_unzip}{any temp folder path for the unzipped archive}

\item{...}{}
}
\value{
ds the extracted data set
}
\description{
Read zip-archive
}
\examples{
\dontrun{
  read_zip()
}
}
