% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_sf_utils.R
\name{sf_latlon_helpers}
\alias{sf_latlon_helpers}
\alias{cast_latlon_to_pts}
\alias{cast_pts_to_ls}
\alias{cast_latlon_to_ls}
\alias{cast_pts_to_latlon}
\title{Utility functions to cast lat/lon data frame/tibble to sf-points and/or linestring}
\usage{
cast_latlon_to_pts(.df, .crs = 4326, .drop_coord = TRUE)

cast_pts_to_ls(.pts_sf, .group_var)

cast_latlon_to_ls(.df, .crs = 4326, .drop_coord = TRUE, ...)

cast_pts_to_latlon(
  .df_pts,
  lat_var = LAT,
  lon_var = LON,
  .crs = 4326,
  .drop_geometry = TRUE,
  ...
)
}
\arguments{
\item{.df}{dataframe/tibble of lat/lon positions}

\item{.crs}{coordinate reference system, e.g. WG84 crs - 4326}

\item{.drop_coord}{remove (or keep) lat/lon columns (default: TRUE := remove, FALSE := keep)}

\item{...}{optional parameters}

\item{.df_pts}{dataframe of POINTS}

\item{.drop_geometry}{flag whether to drop geometry}
}
\value{
utility functions

pts_sf a POINT sf-object

sf linestring

sf linestring

df with LAT/LON coordinates
}
\description{
Utility functions to cast lat/lon data frame/tibble to sf-points and/or linestring
}
\examples{
\donotrun{
cast_latlon_to_pts(adsb_df)
}
}
